moved package to slack_grid
