+++ services/slack_grid/parse_test.go
@@ -1,4 +1,4 @@
-package slack_bulk
+package slack_grid
 
 import (
 	"archive/zip"
